Convert Flow to Bleeding and update values
- Rename Flow Level to Bleeding
- Add new Bleeding values
- Make Bleeding a multiple entry field
- Use FlexboxLayout for displaying Bleeding values
- Upgrade DayData table

+++ app/src/main/kotlin/com/myperioddataismine/DayData.kt
@@ -5,12 +5,12 @@ import java.util.Calendar
 
 class DayData {
     var date: Calendar = Calendar.getInstance()
-    var flowLevel: Int = 0
+    var bleeding: Int = 0
     var moods: Int = 0
     var symptoms: Int = 0
 
     enum class Field {
-        FlowLevel,
+        Bleeding,
         Moods,
         Symptoms
     }
@@ -21,12 +21,15 @@ class DayData {
     }
 
     companion object: Table {
-        val flowLevelValues = arrayOf (
+        val bleedingValues = arrayOf (
             R.string.none to R.drawable.none,
+            R.string.spotting to R.drawable.spotting,
             R.string.light to R.drawable.drop1,
             R.string.moderate to R.drawable.drop2,
             R.string.heavy to R.drawable.drop3,
-            R.string.omg to R.drawable.drop4
+            R.string.small_clots to R.drawable.small_clots,
+            R.string.big_clots to R.drawable.big_clots,
+            R.string.omg to R.drawable.omg
         )
 
         val moodValues = arrayOf(
@@ -57,16 +60,18 @@ class DayData {
         // Table columns
         private const val ID: String = "_id"
         private const val DATE: String = "Date"
-        private const val FLOW_LEVEL: String = "Flow Level"
+        private const val BLEEDING: String = "Bleeding"
         private const val MOOD: String = "Mood"
         private const val SYMPTOMS: String = "Symptoms"
+        // V2 table columns
+        private const val FLOW_LEVEL: String = "Flow Level"
 
         override fun createTable(db: SQLiteDatabase) {
             db.execSQL(
                 "CREATE TABLE \"$PERIOD_DATA\" (" +
                     "\"$ID\" INTEGER PRIMARY KEY AUTOINCREMENT, " +
                     "\"$DATE\" INTEGER NOT NULL UNIQUE, " +
-                    "\"$FLOW_LEVEL\" INTEGER NOT NULL, " +
+                    "\"$BLEEDING\" INTEGER NOT NULL, " +
                     "\"$MOOD\" INTEGER NOT NULL, " +
                     "\"$SYMPTOMS\" INTEGER NOT NULL" +
                 ");"
@@ -74,10 +79,23 @@ class DayData {
         }
 
         override fun upgradeTable(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
-            db.execSQL(
-                "DROP TABLE IF EXISTS \"$PERIOD_DATA\";"
-            )
-            createTable(db)
+            var currentVersion = oldVersion
+            while (currentVersion < newVersion) {
+                when (currentVersion) {
+                    2 -> {
+                        upgradeTableV2toV3(db)
+                        currentVersion++
+                    }
+                    else -> {
+                        println("Failed to upgrade database from $oldVersion to $newVersion")
+                        println("Creating a new database")
+                        db.execSQL(
+                            "DROP TABLE IF EXISTS \"$PERIOD_DATA\";"
+                        )
+                        createTable(db)
+                    }
+                }
+            }
         }
 
         fun get(db:SQLiteDatabase, date: Calendar): DayData {
@@ -85,7 +103,7 @@ class DayData {
             dayData.date = date.clone() as Calendar
             val cursor = db.query(
                 "SELECT " +
-                "\"$FLOW_LEVEL\", " +
+                "\"$BLEEDING\", " +
                 "\"$MOOD\", " +
                 "\"$SYMPTOMS\" " +
                 "FROM \"$PERIOD_DATA\" " +
@@ -93,7 +111,7 @@ class DayData {
             )
             cursor.moveToFirst()
             if (cursor.isAfterLast) return dayData
-            dayData.flowLevel = cursor.getInt(0)
+            dayData.bleeding = cursor.getInt(0)
             dayData.moods = cursor.getInt(1)
             dayData.symptoms = cursor.getInt(2)
             return dayData
@@ -103,13 +121,13 @@ class DayData {
             db.execSQL(
                 "INSERT OR REPLACE INTO \"$PERIOD_DATA\" (" +
                     "\"$DATE\", " +
-                    "\"$FLOW_LEVEL\", " +
+                    "\"$BLEEDING\", " +
                     "\"$MOOD\", " +
                     "\"$SYMPTOMS\"" +
                 ") " +
-                "VALUES(" +
+                "VALUES (" +
                     "'${intForDate(dayData.date)}', " +
-                    "${dayData.flowLevel}, " +
+                    "${dayData.bleeding}, " +
                     "${dayData.moods}, " +
                     "${dayData.symptoms}" +
                 ");"
@@ -122,6 +140,25 @@ class DayData {
             val day = date.get(Calendar.DAY_OF_MONTH)
             val result = year * 10000 + month * 100 + day
             return result
+        }
+
+        private fun upgradeTableV2toV3(db: SQLiteDatabase) {
+            db.execSQL("ALTER TABLE \"$PERIOD_DATA\" RENAME TO \"Old $PERIOD_DATA\"")
+            createTable(db)
+            db.execSQL(
+                "INSERT INTO \"$PERIOD_DATA\" (" +
+                    "\"$DATE\", " +
+                    "\"$BLEEDING\", " +
+                    "\"$MOOD\", " +
+                    "\"$SYMPTOMS\"" +
+                ") SELECT " +
+                "\"$DATE\", " +
+                "1 << \"$FLOW_LEVEL\", " +
+                "\"$MOOD\", " +
+                "\"$SYMPTOMS\" " +
+                "FROM \"Old $PERIOD_DATA\";"
+            )
+            db.execSQL("DROP TABLE \"Old $PERIOD_DATA\"")
         }
     }
 }
